Allow either clang or LLVM to trigger builds
This is required to support: https://codereview.chromium.org/1577943002/

Eventually we'll want other repositories to also be able to trigger builds. Maybe the lastest and lkgr files should also contain all the hashes that they've checked out, not just the one from BUILDBOT_REVISION. With that the build could use lkgr to only trigger on one repository, and keep all other repositories at lkgr (so they don't all float).

The output is now json, and contains information on all the repos as well as the build number and what schedule triggered the build. The archives are now named based on build number.

The following binaryen patch teaches binaryen this new format:
  https://github.com/WebAssembly/binaryen/commit/9a3a2cb975ab171c4f23c5bf5e938db940c707d4