Limit the number of characters that are compared per output
All failing outputs are compared with each other. That takes ~30s for the GCC torture tests versus ~12s when truncating the output to 1024 characters. Nothing useful is likely to be beyond this.

SequenceMatcher also supports an isjunk predicate, but we'd need to tune it for each output and that seems like too much work.