Fix some missing libc functionality
This implements a super hacky libc skeleton in JavaScript. It's not pretty. Don't do this. But hey it works and we can use it now. The goal for the near future will instead be to use musl, but this hack should get us off the ground easily. Don't look too closely lest your eyes burn.