Use Binaryen's wasm-as to encode wasm binaries instead of sexpr-wasm (#53)
Emscripten's default behavior when generating .js files with wasm is now
to use wasm-as to encode a binary (previously it generated a wast
file). Since the binary 11 format, the various tools and engines have
had good agreement on the binary format and have synchronized fairly
well, so it should be safe to do this. When the format changes again we
can switch back to the previous behavior if necessary.