Add support for running emscripten test suite
Run the emscripten test suite with the wasm backend, using the
wasm-shell (as opposed to native wasm) for now.
Also switch to Node as the JS_ENGINES used to run the test (previously
this value was V8 but unused).
Also remove the attempt to use a relative path in the emscripten config
file and instead generate one with an absolute path at install time.