Make Filter.Check not throw on unknown targets
Throwing on unknown targets only makes sense when a list of possible
targets is given (e.g. when calling Apply), not when using bare
Check (since the filter does not necessarily contain all known targets
itself).

This refactors the actual exclusion logic into check, and implements
Apply in terms of that.