Create emcc/em++ wrappers in toolchain bin directory
These wrappers set/override EM_CONFIG meaning that the user's emscripten config
won't interfere with the toolchain. They also set EMCC_SKIP_SANITY_CHECK so that
emscripten won't try to write to the toolchain directory (which is not writable
in the packaged version).

Also, add /etc/alternatives for the core set of binaries in the debian package
so they are accessible via /usr/bin.

Also, fix the pyc cleanup command, and don't remove the third_party directory
(it is needed for closure-compiler.jar at least).