Use git bash from depot_tools for LLVM regression tests
Several of clang's regression tests use some nontrivial unix utilities
and/or bash features. We've been using the gnuwin32 ports of these
utilities, but for reasons I don't fully understand yet, this
reliably causes some issues with temp files created by tests on
windows 7 (windows 10 seems fine). The git that comes with
depot_tools has an msys bash with msys versions of these utilities
that are sufficient for running the LLVM tests, so switch to that
for now.