Implement test attributes for known failure files (#181)
Augment the syntax used in test exclusion ('known failure') files to use attributes for torture test run configurations and exclusions. This allows most of the test exclusions to be combined into one file, where they can use common explanations, and where it's easier to see which combinations of conditions cause the failures.
This PR combines the bare JS engine and emscripten test exclusions into a single file, but not the spec interpreter (it has a large set and most failures are just due to missing imports; we may try it in the future) or the Binaryen shell (it lives in a different repo), or the compile failures (which are of course disjoint from the runtime failures).