Fix download link for "Archive binaries" step
This is partial fix for #197.  For steps that upload
more than one archive only one link is shown in the
buildbot GUI.  This change makes uploads the binaries
last in the step which fixes the link.

Secondly, remove the build number from the filenames
since it is already present in the directory name.