Make JSC build failure not abort the whole process (#223)
* Make JSC build failure not abort the whole process

Currently JSC build is using cmake build, which is not the official
build, and this makes JSC build flaky. This makes JSC build failure
not stop the whole waterfall process so that we can proceed to the
test stage.
Also, the summary will print warning-only failures as well for the
information.