Make waterfall test multiple optimization levels
Currently bare tests are only compiled with -O0, whereas emcc/binaryen
tests are only compiled with -O2. This makes bare tests compiled with
both -O0 and -O2, and emcc/binaryen tests with both -O0 and -O3.
(The reason for not -O2 again but -O3 is people tend to use -O3 a lot in
a real usecase and test emcc with a different set of flags will increase
test coverage.)