Add untriaged cases to the expectations (#232)
As we added multiple compilation optimization flags to waterfall, and a
and as we are planning to change many tests to non-warn-only, we get
many unexpected successes and failures. This adds those cases to the
failure expectations as a temporary measure to make waterfall green
again. They will be triaged and re-categorized soon.

As we are not planning to change wasm-shell tests to non-warn-only at
the moment, this PR does not address its unexpected successes and
failures.