Disable sanity check for emscripten build
Each invocation of emcc does sanity check for the current emscripten
cache directory, whose default location is `~/.emscripten_cache`. It
compares the last modification of the configuration file
(`emscripten_config` for asmjs and `emscripten_config_vanilla` for wasm)
with that of the sanity file (`emscripten_config_sanity` for asmjs and
`emscripten_config_vanilla_sanity_wasm` for wasm), and if the
configuration file was modified after the sanity file, deletes the cache
directory and rewrites the sanity file.
https://github.com/kripken/emscripten/blob/a243a941d911698a703fd4e5210a9dbc24bc4e96/tools/shared.py#L512-L533
https://github.com/kripken/emscripten/blob/a243a941d911698a703fd4e5210a9dbc24bc4e96/tools/shared.py#L578-L582

The problem is, emcc thinks the cache is invalid when
`sanity file modification time <= config file modification time`, which
means the cache will be deleted even if two timestamps are equal.
https://github.com/kripken/emscripten/blob/a243a941d911698a703fd4e5210a9dbc24bc4e96/tools/shared.py#L520
This does not happen usually in Linux, but in Mac OS X and Windows, [the
file modification timestamps have only 1 or 2 seconds
resolution](https://docs.python.org/3/library/os.html#os.stat). And
the system library creation by embuilder involves several invocation of
emcc: emcc is invoked per every library (.a or .bc) to build, each of
which also launches multiple emcc processes for each .c file that
belongs to the library. This means it is possible that the cache is
deemed invalid and deleted in the first emcc invocation, but child emcc
processes wlll also delete the cache again, which would then contain
half-built libraries, if less than 1 seconds has elapsed after the the
first deletion.

Modifying the sanity check equation from `<=` to `<` so that the cache
will only be deleted when
`sanity file modification time < config file modification time` dos not
seem to be a good solution, because it is also possible that in other
case it is possible that config file creation is followed by sanity file
creation, which warranties the regeneration of sanity file.

This patch just disables checking of sanity file timestamp. The
emscripten cache directory is deleted in the beginning of the function
anyway, so it's ok not to check it for the first emcc call as well.