Debugging hack for Mac. Will revert soon.
In emscripten testsuite, test_openjpeg and test_the_bullet fail only for
Mac, when they run 'configure' command. I can't reproduce the failure in
my local machine. This makes the log to print configure output, and it
only executes a single test case for this.
Will revert within a hour.