Add ability to completely exclude torture tests from test pipeline (#308)
The test in question has UB and is optimized into an infinite loop. On mac and
linux the python rlimit mechanism is used to kill the program after timeout, but
there's no equivalent functionality for Windows currently. So we just exclude
the test entirely (this also makes everything run faster on mac and linux since
we aren't waiting uselessly for the timeout).