Add jsvu step
This patch adds a step that installs [jsvu](https://github.com/GoogleChromeLabs/jsvu). For now it only does so on Windows (where jsvu then installs a `chakra` binary) and Mac (where jsvu then installs a `javascriptcore` binary), but this can easily be extended in the future.

Hopefully this can replace the complex build systems for these engines and reduce the maintenance cost of this project.