Add jsvu step (#325)
This patch adds a step that installs [jsvu](https://github.com/GoogleChromeLabs/jsvu). For now it only does so Mac (where jsvu then installs a `javascriptcore` binary), but this can easily be extended in the future.

Hopefully this can replace the complex build systems for these engines and reduce the maintenance cost of this project.