Update toolchain files and CMake builds for more correct feature detection (#341)
Refactor the toolchain files int one that's easy to include (i.e. not buried in a weird directory hierarchy) and one that's in the location where CMake expects to find platform files (and avoids warnings).

It also gives the CMake configuration a less-generic name (WACK: the WebAssembly Clang Kit) and makes it more clear that the toolchain built by the waterfall is just one particular experimental configuration but not the only possible one. It's also easier to refer to a particular name instead always saying "the toolchain that the waterfall builds".