Remove s2wasm testing from the waterfall
This is a fairly large change since many part of the waterfall
were using s2wasm or its output.

We now only compiler the toture tests once, to wasm object files.
However we link them twice, once with -nostdlib and once without.
This is a because we can't run binaries linked with musl in the
spec interpreter.