Remove wasm-shell testing of torture tests
This is part of the move to remove s2wasm.  When we do this
we will only producing wasm binaries from the toolchain and
not wast/wat files.

wasm-shell only supports running on the .wast files (it designed
to run the spec tests).  (It you pass it wasm file it actually just
returns 1 since wasm files are all empty strings in C!)

The wasm-shell tests have been failing for a while but we have
the in warn only mode, which is another reason I don't think we
care about coverage here.