Update expectations after llvm upstream change.
As of https://reviews.llvm.org/D57909 lld will no longer
generate invalid modules when function signatures mistmatch.

Instead is tries to figure out the actual functions signature.
In this case it fails to do so because the disagreement is between
two imports and there is no actual definition of `abort`.