Move chromium-clang script source and prebuilt clang out of their separate directory (#489)
Instead put both of them in the root of the working directory. (The extraction path of clang is hardcoded in the script, so keep the 'tools/clang' directory structure).
Also make the prebuilt tools use a separately-configurable working directory.
Lining everything up this way makes things simpler when using DEPS to manage the sources.