Use V8's checkout of the chrome build repo instead of managing our own
Currently we have our own checkout of chrome's build repo (to get the
host clang and the windows SDK) but use V8's checkout to actually
configure it because of a quirk in the directory paths they assume.

This switches to using V8's version consistently.
This has the advantage that we don't need to manage our own checkout
or DEPS entry and don't duplicate V8 logic.