Avoid libtinfo specifically on linux, where we are seeing issues (#541)
As not having it removes colors on some systems. Oddly on one machine I didn't see this, but testing on another, I do - all the colors are gone.

The eventual solution may be to statically link libtinfo. Meanwhile with it we get errors on some linuxes but proper colors, and without it we avoid linux errors but have no colors. With this PR at least the loss of colors is just on linux.

See

emscripten-core/emsdk#252

#539

#540