Install libclang_shared.so.* into wasm-install/lib
After
https://github.com/llvm/llvm-project/commit/2e97d2aa1bd313350e76be64299f19172a346bf9,
`CLANG_LINK_CLANG_DYLIB` is `ON` whenever `LLVM_LINK_LLVM_DYLIB` is
`ON`. This requires `libclang_shared.so` to be installed in
`wasm-install/lib`, but `-DLLVM_INSTALL_TOOLCHAIN_ONLY=ON` currently
disables its installation. This adds `libclang_shared.so` to `extra_lib`
so it can be installed. This fixes the current emscripten build step
failure in the waterfall.