Update expectations on failing asan tests (#576)
We don't run sanitizers on them, so all sanitizer tests (ubsan, asan,
tsan, ...) may start to fail at any point, and we shouldn't care about
the results. For now I update the expectation to match the current
behavior, but if this happens frequently enough, we may need to fix the
runner script so that we can skip those tests or something.

The two asan tests started to fail after CraneStation/wasi-libc#112, just for
reference.