Add build target for LLVM test dependencies
These are regular host build targets which are required for running the LLVM
regression tests (make check) but which do not build by default from the 'all'
target. An extra build rule is used (instead of bundling it with the test
running) because the emscripten-releases shuts down goma before running
tests.