Trigger LTO builds on locally-modified version file
Currently LTO builds are only triggered when the most recent commit modifies
the version file. But on try jobs, the files changed by the CL are not
committed; they are just modified in the index. So this change causes
LTO builds to also be triggered when the version file is modified
according to 'git status'.