Pass correct SDK path to build's setup_toolchain utility
When using Chrome's hermetic toolchain, we also use a build infra utility to
set up the environment variables that VS-compatible toolchains expect.
Until recently one of the arguments to that script was irrelevant for us,
but it recently became required (now the first argument points to the
root of the toolchain package, and the 2nd sdk_path must be a subdir).